requirements.txt
azure-core==1.32.0
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
Flask==3.1.0
gitdb==4.0.11
GitPython==3.1.41
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msal==1.31.1
msal-extensions==1.2.0
portalocker==2.10.1
pycparser==2.22
PyJWT==2.10.0
requests==2.32.3
setuptools==69.0.3
six==1.16.0
smmap==5.0.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wheel==0.44.0
